EID-918: Move exceptions into an exceptions package
Authors: @kgarwood, @felixmarcusmillne
