+++ src/main/java/uk/gov/ida/metadataaggregator/MetadataAggregator.java
@@ -5,9 +5,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import uk.gov.ida.metadataaggregator.configuration.MetadataSourceConfiguration;
 import uk.gov.ida.metadataaggregator.metadatasource.CountryMetadataResolver;
-import uk.gov.ida.metadataaggregator.metadatasource.MetadataSourceException;
+import uk.gov.ida.metadataaggregator.exceptions.MetadataSourceException;
 import uk.gov.ida.metadataaggregator.metadatastore.MetadataStore;
-import uk.gov.ida.metadataaggregator.metadatastore.MetadataStoreException;
+import uk.gov.ida.metadataaggregator.exceptions.MetadataStoreException;
 
 import java.net.URL;
 import java.util.ArrayList;
