+++ src/main/java/uk/gov/ida/metadataaggregator/MetadataAggregatorModule.java
@@ -6,7 +6,7 @@ import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 import com.google.inject.name.Named;
-import uk.gov.ida.metadataaggregator.configuration.ConfigSourceException;
+import uk.gov.ida.metadataaggregator.exceptions.ConfigSourceException;
 import uk.gov.ida.metadataaggregator.configuration.MetadataAggregatorConfiguration;
 import uk.gov.ida.metadataaggregator.configuration.MetadataSourceConfiguration;
 import uk.gov.ida.metadataaggregator.configuration.MetadataSourceConfigurationLoader;
