+++ src/main/java/uk/gov/ida/metadataaggregator/exceptions/ConfigSourceException.java
@@ -1,4 +1,4 @@
-package uk.gov.ida.metadataaggregator.configuration;
+package uk.gov.ida.metadataaggregator.exceptions;
 
 public class ConfigSourceException extends Exception {
     public ConfigSourceException(String s, Throwable e) {
