+++ src/main/java/uk/gov/ida/metadataaggregator/exceptions/EnvironmentVariableException.java
@@ -1,4 +1,4 @@
-package uk.gov.ida.metadataaggregator;
+package uk.gov.ida.metadataaggregator.exceptions;
 
 public class EnvironmentVariableException extends Exception {
     public EnvironmentVariableException(String message) {
