+++ src/main/java/uk/gov/ida/metadataaggregator/exceptions/MetadataStoreException.java
@@ -1,4 +1,4 @@
-package uk.gov.ida.metadataaggregator.metadatastore;
+package uk.gov.ida.metadataaggregator.exceptions;
 
 public class MetadataStoreException extends Exception {
     public MetadataStoreException(String s, Throwable e) {
